finalize parser tool
