+++ Sources/TelemetryParser/Helpers/RegexExtractor.swift
